+++ backend/src/modules/classes.py
@@ -2,24 +2,56 @@
 ChatGPT related classes and response variables
 """
 # pylint: disable=line-too-long,too-few-public-methods,no-name-in-module
-from typing import Optional
+from typing import Optional, List
 from pydantic import BaseModel, Field
 
 
 class ChatJson(BaseModel):
     """
-    ChatGPT chat endpoint expected json input styling
+    Open AI completion endpoint expected json input styling
     """
 
     model: str = Field(
-        ...,
-        description="The name of the model you want to use",
+        ..., # Pydantic notation for a required Field https://docs.pydantic.dev/dev/concepts/models/#required-fields
+        description="(Required if Open AI) The name of the model you want to use",
         example="gpt-3.5-turbo",
     )
     messages: list = Field(
         ...,
+        description="(Required if Open AI) List of messages with roles associated",
+        example="""[{"role": "user", "content": "Hello!"}]""",
+    )
+    messages: Optional[list] = Field(
+        ...,
+        description="(Required if Open AI) List of messages with roles associated",
+        example="""[{"role": "user", "content": "Hello!"}]""",
+    )
+    root_id: str = Field(
+        None,
+        description="The root UUID of the conversation, if applicable. This is how you connect conversations",
+    )
+    convo_title: Optional[str] = Field(
+        None,
+        description="The title of the conversation, if applicable. This is needed if you change convo titles",
+    )
+
+class ChatSagemakerJson(BaseModel):
+    """
+    Customized class for sagemaker based generative ai endpoint. Modify as needed for which ever llm model you use.
+    """
+
+    inputs: List[dict] = Field(
+        ...,
         description="list of messages with roles associated",
         example="""[{"role": "user", "content": "Hello!"}]""",
+    )
+    parameters: dict = Field(
+        description="kwargs for the llama model. See more here: https://huggingface.co/docs/transformers/main_classes/text_generation#transformers.GenerationConfig",
+        example= {
+		"max_new_tokens": 512,
+		"top_p": 0.9,
+		"temperature": 0.6
+	    }
     )
     root_id: Optional[str] = Field(
         None,
@@ -30,6 +62,40 @@ class ChatJson(BaseModel):
         description="The title of the conversation, if applicable. This is needed if you change convo titles",
     )
 
+class ChatBedrockJson(BaseModel):
+    """
+    Customized class for bedrock based generative ai endpoint. Modify as needed for which ever llm model you use.
+    Additional Documentation here: https://docs.aws.amazon.com/bedrock/latest/userguide/api-setup.html
+    """
+
+    modelId: str = Field(
+        ...,
+        description="(Required if AWS Bedrock) The name of the model you want to use. ModelIds listed here: https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html.",
+        example="ai21.j2-mid-v1",
+    )
+    prompt: str = Field(
+        ...,
+        description = "Single prompt to pass to the LLM model",
+        example = "What is the difference between Python and Go?"
+    )
+    maxTokens: Optional[int] = Field(
+        description = "The maximum number of tokens to generate per result. Optional, default = 16. (Source https://docs.ai21.com/reference/j2-complete-ref).",
+        example=200
+    )
+    root_id: Optional[str] = Field(
+        None,
+        description="The root UUID of the conversation, if applicable. This is how you connect conversations",
+    )
+    convo_title: Optional[str] = Field(
+        None,
+        description="The title of the conversation, if applicable. This is needed if you change convo titles",
+    )
+    class Config:
+        """
+        Config to allow open inputs
+        """
+
+        extra = "allow"
 
 class CheckChatJson(BaseModel):
     """
@@ -47,7 +113,6 @@ class CheckChatJson(BaseModel):
         example="""[{"role": "user", "content": "Hello!"}]""",
     )
 
-
 class TitleUpdateJson(BaseModel):
     """
     ChatGPT title update json input styling
@@ -64,7 +129,6 @@ class TitleUpdateJson(BaseModel):
         example="""new title""",
     )
 
-
 class RootUuidJson(BaseModel):
     """
     ChatGPT json input styling which only expects a root_id
@@ -75,7 +139,6 @@ class RootUuidJson(BaseModel):
         description="The root UUID of the conversation, if applicable. This is how you connect conversations",
         example="""xxxxxxxx-xx-xxxx-xxxx-xxxxxx""",
     )
-
 
 class ImgJson(BaseModel):
     """
@@ -91,7 +154,6 @@ class ImgJson(BaseModel):
         description="size of the image you want, three are available: 1024x1024, 512x512, 256x256",
         example="1024x1024",
     )
-
 
 unknown_error = {
     "description": "Unknown error",
